convert fstring rewriter to a plugin
